refactor(indexer): switch to shated logic for processing change set with SQS
