+++ ingest/indexer/service/indexer_streaming_service.go
@@ -17,16 +17,14 @@ var _ storetypes.ABCIListener = (*indexerStreamingService)(nil)
 
 // ind is a streaming service that processes block data and ingests it into the indexer
 type indexerStreamingService struct {
-	writeListeners map[storetypes.StoreKey][]commondomain.WriteListener
-
 	// manages tracking of whether the node is code started
 	coldStartManager indexerdomain.ColdStartManager
-
-	storeKeyMap map[string]storetypes.StoreKey
 
 	client indexerdomain.Publisher
 
 	keepers indexerdomain.Keepers
+
+	blockUpdatesProcessUtils commondomain.BlockUpdateProcessUtilsI
 }
 
 // New creates a new sqsStreamingService.
@@ -34,18 +32,16 @@ type indexerStreamingService struct {
 // sqsIngester is an ingester that ingests the block data into SQS.
 // poolTracker is a tracker that tracks the pools that were changed in the block.
 // nodeStatusChecker is a checker that checks if the node is syncing.
-func New(writeListeners map[storetypes.StoreKey][]commondomain.WriteListener, coldStartManager indexerdomain.ColdStartManager, client indexerdomain.Publisher, storeKeyMap map[string]storetypes.StoreKey, keepers indexerdomain.Keepers) storetypes.ABCIListener {
+func New(blockUpdatesProcessUtils commondomain.BlockUpdateProcessUtilsI, coldStartManager indexerdomain.ColdStartManager, client indexerdomain.Publisher, storeKeyMap map[string]storetypes.StoreKey, keepers indexerdomain.Keepers) storetypes.ABCIListener {
 	return &indexerStreamingService{
-
-		writeListeners: writeListeners,
 
 		coldStartManager: coldStartManager,
 
 		client: client,
 
-		storeKeyMap: storeKeyMap,
-
 		keepers: keepers,
+
+		blockUpdatesProcessUtils: blockUpdatesProcessUtils,
 	}
 }
 
@@ -149,21 +145,11 @@ func (s *indexerStreamingService) ListenCommit(ctx context.Context, res abci.Res
 		// Mark that the initial data has been ingested
 		s.coldStartManager.MarkInitialDataIngested()
 	} else {
-		for _, kv := range changeSet {
-			for _, listener := range s.writeListeners[s.storeKeyMap[kv.StoreKey]] {
-				if err := listener.OnWrite(s.storeKeyMap[kv.StoreKey], kv.Key, kv.Value, kv.Delete); err != nil {
-					return err
-				}
-			}
-		}
+		s.blockUpdatesProcessUtils.SetChangeSet(changeSet)
+		s.blockUpdatesProcessUtils.ProcessBlockChangeSet()
 	}
 
 	return nil
-}
-
-// Listeners implements baseapp.StreamingService.
-func (s *indexerStreamingService) Listeners() map[storetypes.StoreKey][]commondomain.WriteListener {
-	return s.writeListeners
 }
 
 // Stream implements baseapp.StreamingService.
