+++ ingest/sqs/service/blockprocessor/block_process_sqs_factory.go
@@ -12,7 +12,7 @@ type transformAndLoadFunc func(ctx sdk.Context, poolsTrasnformer domain.PoolsTra
 // NewBlockProcessor creates a new block process strategy.
 // If block process strategy manager should push all data, then it will return a full indexer block process strategy.
 // Otherwise, it will return a block updates SQS block process strategy.
-func NewBlockProcessor(blockProcessStrategyManager commondomain.BlockProcessStrategyManager, client domain.SQSGRPClient, poolExtractor commondomain.PoolExtractor, poolsTransformer domain.PoolsTransformer, nodeStatusChecker domain.NodeStatusChecker, blockUpdateProcessUtils domain.BlockUpdateProcessUtilsI) commondomain.BlockProcessor {
+func NewBlockProcessor(blockProcessStrategyManager commondomain.BlockProcessStrategyManager, client domain.SQSGRPClient, poolExtractor commondomain.PoolExtractor, poolsTransformer domain.PoolsTransformer, nodeStatusChecker domain.NodeStatusChecker, blockUpdateProcessUtils commondomain.BlockUpdateProcessUtilsI) commondomain.BlockProcessor {
 	// If true, ingest all the data.
 	if blockProcessStrategyManager.ShouldPushAllData() {
 		return &fullSQSBlockProcessStrategy{
