+++ ingest/sqs/service/sqs_streaming_service.go
@@ -31,7 +31,7 @@ type sqsStreamingService struct {
 
 	nodeStatusChecker domain.NodeStatusChecker
 
-	blockUpdatesProcessUtil domain.BlockUpdateProcessUtilsI
+	blockUpdatesProcessUtil commondomain.BlockUpdateProcessUtilsI
 }
 
 // New creates a new sqsStreamingService.
@@ -39,7 +39,7 @@ type sqsStreamingService struct {
 // sqsIngester is an ingester that ingests the block data into SQS.
 // poolTracker is a tracker that tracks the pools that were changed in the block.
 // nodeStatusChecker is a checker that checks if the node is syncing.
-func New(blockUpdatesProcessUtil domain.BlockUpdateProcessUtilsI, poolsExtractor commondomain.PoolExtractor, poolsTransformer domain.PoolsTransformer, poolTracker domain.BlockPoolUpdateTracker, grpcClient domain.SQSGRPClient, blockProcessStrategyManager commondomain.BlockProcessStrategyManager, nodeStatusChecker domain.NodeStatusChecker) *sqsStreamingService {
+func New(blockUpdatesProcessUtil commondomain.BlockUpdateProcessUtilsI, poolsExtractor commondomain.PoolExtractor, poolsTransformer domain.PoolsTransformer, poolTracker domain.BlockPoolUpdateTracker, grpcClient domain.SQSGRPClient, blockProcessStrategyManager commondomain.BlockProcessStrategyManager, nodeStatusChecker domain.NodeStatusChecker) *sqsStreamingService {
 	return &sqsStreamingService{
 		blockUpdatesProcessUtil:     blockUpdatesProcessUtil,
 		poolsExtractor:              poolsExtractor,
